further improvement of makesymlinks.sh

+++ makesymlinks.sh
@@ -23,12 +23,18 @@ echo -n "Changing to the $dir directory ..."
 cd $dir
 echo "done"
 
+echo -n "Pulling dotfiles repo..."
+git pull
+echo "done"
+
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+  echo "Moving any existing dotfiles from ~ to $olddir"
+  if [ -f ~/.$file ];then
+      mv ~/.$file ~/dotfiles_old/
+  fi
+  echo "Creating symlink to $file in home directory."
+  ln -s $dir/$file ~/.$file
 done
 
 install_zsh () {
@@ -37,6 +43,7 @@ if [ -f /bin/zsh -o -f /usr/bin/zsh ]; then
     # Clone my oh-my-zsh repository from GitHub only if it isn't already present
     if [[ ! -d $dir/oh-my-zsh/ ]]; then
         git clone http://github.com/robbyrussell/oh-my-zsh.git
+    fi
     # Set the default shell to zsh if it isn't currently set to zsh
     if [[ ! $(echo $SHELL) == $(which zsh) ]]; then
         chsh -s $(which zsh)
